# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/widget/sessions": {
      "post": {
        "description": "Validates the widget key together with the request `Origin` (the key must list that origin; a call without an Origin header is refused), checks that the tenant may chat, creates or finds the visitor (`web_<visitorId>`), creates the conversation (or resumes the open one of this visitor: call again with the same `visitorId` to get a fresh token before the 15 minutes end) and returns the widget token. A suspended or closed tenant, a plan without chat, or an engine that is down answers **200 with `status: \"blocked\"`** and a `fallback` text instead of an error; a tenant at its conversation limit gets `status: \"limited\"`. Errors: 401 WIDGET_KEY_INVALID, 403 ORIGIN_NOT_ALLOWED, 429 TOO_MANY_REQUESTS (with `Retry-After`).",
        "operationId": "WidgetController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWidgetSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetSession"
                }
              }
            }
          }
        },
        "summary": "Start or refresh a chat session (public, widget key + Origin)",
        "tags": [
          "widget"
        ]
      }
    },
    "/v1/widget/messages": {
      "post": {
        "description": "The response is a **server-sent event stream** (`text/event-stream`), read with `fetch` and a stream reader (a POST cannot use `EventSource`). Events, in order: `accepted` (message stored), then `token` (zero or more; concatenate `text`), optionally `escalated` (a human was asked to take over), and finally either `done` or `fallback` (the AI did not answer: `service_unavailable`, `limit_reached` or `ai_unavailable`; the customer is never left without a text). `error` (`CONVERSATION_NOT_FOUND`, `CONVERSATION_RESOLVED`) means call POST /v1/widget/sessions for a new conversation. Failures before the stream starts are ordinary JSON errors: 400 MESSAGE_TOO_LONG (over 2,000 characters), 401 WIDGET_TOKEN_EXPIRED, 403 ORIGIN_NOT_ALLOWED, 404 CONVERSATION_NOT_FOUND, 429. Each event is `event: <name>` + `data: <json>` + blank line; the payload schemas are listed under `Sse*Event`.",
        "operationId": "WidgetController_sendMessage",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A unique id (8-100 characters of A-Z a-z 0-9 _ -) per message. Sending the same key again does not store the message twice and does not call the model again.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendWidgetMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "text/event-stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "event: accepted\ndata: {\"messageId\":\"…\"}\n\nevent: token\ndata: {\"text\":\"Our \"}\n\nevent: done\ndata: {\"messageId\":\"…\",\"aiReply\":true,\"conversationStatus\":\"active\"}\n\n"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a customer message and stream the reply (widget token)",
        "tags": [
          "widget"
        ]
      }
    },
    "/v1/widget/events": {
      "get": {
        "description": "A `text/event-stream` for what happens while the customer is not sending: a staff reply, the system lines (\"an agent joined\") and status changes. Like the message route it needs the `Authorization` header, so read it with `fetch` and a stream reader, not `EventSource`; the key and the Origin are checked as on every widget call.\n\nEvents: `message` `{id, authorType: \"human\" | \"system\", content, contentKey, createdAt}` (a staff reply has text and no identity; a system line has an empty `content` and a `contentKey` of the widget namespace such as `agent.joined`), `status` `{status: \"active\" | \"escalated\" | \"human_active\" | \"resolved\"}`. The AI's own replies and the customer's messages are NOT sent here (they come through the message stream). Show a `message` once, by its `id` (it is also in the history).\n\nControl events: `ready`, `resync` (your `Last-Event-ID` is too old: re-read `GET /v1/widget/conversation`), `closed` `{reason}` (`expired` when the 15-minute token ends: refresh the session and reconnect with the new token and the last event id; `revoked`; `limit` when a fourth stream of the same conversation replaced this one). A `: ping` comment arrives every 25 seconds. Every data event has an `id` for `Last-Event-ID` (the last 100 are kept; the history route is the fallback).\n\nErrors before the stream starts are ordinary JSON: 401, 403 ORIGIN_NOT_ALLOWED, 403 TENANT_SUSPENDED, 404 CONVERSATION_NOT_FOUND, 429.",
        "operationId": "WidgetController_events",
        "parameters": [],
        "responses": {
          "200": {
            "description": "text/event-stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "retry: 3000\n\nevent: ready\ndata: {\"heartbeatSeconds\":25}\n\nid: 9f2a-1\nevent: message\ndata: {\"id\":\"…\",\"authorType\":\"human\",\"content\":\"Hello, I can help\",\"contentKey\":null,\"createdAt\":\"…\"}\n\n"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Live events of the visitor's conversation (widget token)",
        "tags": [
          "widget"
        ]
      }
    },
    "/v1/widget/conversation": {
      "get": {
        "description": "Oldest first, paged. Returns only the conversation named in the token; internal (tool) messages and staff identities are left out. 503 ENGINE_UNAVAILABLE when the engine cannot be reached.",
        "operationId": "WidgetController_history",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetConversation"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The conversation history of the visitor (widget token)",
        "tags": [
          "widget"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-tenant support platform: backend core",
    "description": "Tenants, staff authentication, invitations and roles, audit log, translations, end customers. Errors always have the shape `{ statusCode, code, message }`; lists are `{ data, total, skip, take }`.",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWidgetSessionDto": {
        "type": "object",
        "properties": {
          "widgetKey": {
            "type": "string",
            "example": "wk_AbCd1234…",
            "description": "The tenant's public widget key (from the dashboard). It is an identifier, not a secret: it only works from the origins listed for it."
          },
          "visitorId": {
            "type": "string",
            "example": "3f9c1b7e-52aa-4d0e-9a41-6f1d2b8c7e10",
            "description": "A random id the widget creates once (crypto.randomUUID()) and keeps in localStorage. 16 to 64 characters of A-Z a-z 0-9 _ -. Whoever knows it can resume that visitor's chat, so it must be unguessable and is never shown or logged."
          },
          "locale": {
            "type": "string",
            "example": "ur",
            "description": "The visitor's language (the browser's, e.g. `ur-PK`). Falls back to the tenant default, then `en`."
          }
        },
        "required": [
          "widgetKey",
          "visitorId"
        ]
      },
      "WidgetFallback": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "service_unavailable",
              "limit_reached",
              "ai_unavailable"
            ],
            "description": "`service_unavailable`: the chat is switched off for this tenant (suspended, closed, no chat in the plan). `limit_reached`: the plan limit was reached and a human takes over. `ai_unavailable`: the assistant is down or too slow and a human takes over. The customer never sees billing details."
          },
          "message": {
            "type": "string",
            "description": "Text to show the customer, already in their language: the tenant's own fallback message when it has one (Phase 5), else the default from the `widget` namespace."
          }
        },
        "required": [
          "reason",
          "message"
        ]
      },
      "WidgetSession": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "limited",
              "blocked"
            ],
            "description": "`ready`: chat normally. `limited`: the conversation exists but the AI will not answer it (plan limit); messages reach a human; `fallback` says what to show. `blocked`: no chat possible, there is no token; show `fallback`."
          },
          "token": {
            "type": "string",
            "description": "Widget token (Authorization: Bearer) for the other widget routes. Absent when `blocked`. Valid 15 minutes; call this endpoint again with the same visitorId to get a new one."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresInSeconds": {
            "type": "number",
            "example": 900
          },
          "conversationId": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "example": "ur",
            "description": "Language to render the widget in."
          },
          "defaultLocale": {
            "type": "string",
            "example": "en",
            "description": "The tenant's default language."
          },
          "greeting": {
            "type": "string",
            "description": "Greeting to show first (in `locale`)."
          },
          "personaName": {
            "type": "string",
            "description": "Name of the assistant."
          },
          "poweredBy": {
            "type": "boolean",
            "description": "Whether the widget must show the \"Powered by\" label (Free plan)."
          },
          "fallback": {
            "$ref": "#/components/schemas/WidgetFallback"
          }
        },
        "required": [
          "status",
          "locale",
          "defaultLocale",
          "poweredBy"
        ]
      },
      "SendWidgetMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "What are your opening hours?",
            "maxLength": 2000,
            "description": "The customer's text, at most 2,000 characters (longer: 400 MESSAGE_TOO_LONG). Surrounding whitespace is trimmed."
          }
        },
        "required": [
          "content"
        ]
      },
      "WidgetHistoryMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "authorType": {
            "type": "string",
            "enum": [
              "customer",
              "ai",
              "human",
              "system"
            ]
          },
          "content": {
            "type": "string"
          },
          "contentKey": {
            "type": "string",
            "nullable": true,
            "description": "Translation key (`widget` namespace) for system lines such as `agent.joined`; then `content` is empty."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "authorType",
          "content",
          "createdAt"
        ]
      },
      "WidgetConversation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "escalated",
              "human_active",
              "resolved"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WidgetHistoryMessage"
            }
          },
          "total": {
            "type": "number"
          },
          "skip": {
            "type": "number"
          },
          "take": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "status",
          "data",
          "total",
          "skip",
          "take"
        ]
      },
...
    }
  }
}